Changed "RELEASE" to "20.04.0-PATCH1"
This change is to accomodate maven builds in GDC.

If "RELEASE" was used, maven will try to connect to hub.ecquaria.com,
which will fail the whole build because it is unable to determine the
correct version to use.

Maven will also highlight that using "RELEASE", or "LATEST" have been
deprecated and may cause instability (in this case it already does).